[refactor] 사용자 조회 URL /friends -> /users 변경
