+++ src/main/java/com/planpal/demo/converter/UserConverter.java
@@ -1,6 +1,7 @@
 package com.planpal.demo.converter;
 
 import com.planpal.demo.domain.User;
+import com.planpal.demo.web.dto.UserResponseDto.GetResultDto;
 import com.planpal.demo.web.dto.UserResponseDto.SignUpResultDto;
 import com.planpal.demo.web.dto.kakao.KakaoUserInfoDto;
 
@@ -14,5 +15,12 @@ public class UserConverter {
 
     public static SignUpResultDto toSignUpResultDto(String accessToken) {
         return new SignUpResultDto(accessToken);
+    }
+
+    public static GetResultDto toGetResultDto(User user) {
+        return new GetResultDto(
+                user.getId(),
+                user.getNickname(),
+                user.getTagId());
     }
 }
